pyproject.toml
[tool.poetry]
name = "catflow-inference"
version = "0.2.0"
description = "Inference service for an object recognition pipeline"
authors = ["Ian Kilgore <iank@iank.org>"]
license = "MIT"
readme = "README.md"
packages = [{include = "catflow_inference"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.97.0"
numpy = "^1.24.3"
Pillow = "^9.5.0"
uvicorn = "^0.22.0"
torch = "2.0.1+cpu"
torchvision = "0.15.2+cpu"
httpx = "^0.24.1"
pandas = "^2.0.2"
psutil = "^5.9.5"
tqdm = "^4.65.0"
pyyaml = "^6.0"
ultralytics = "^8.0.117"
opencv-python-headless = "^4.7.0.72"
frameextractor = {git = "https://github.com/iank/catflow-frameextractor", rev = "v0.5.5"}
aiofiles = "^23.1.0"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.3.2"
ruff = "^0.0.272"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/"
priority = "supplemental"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = [
    # Fail on any warnings that are not explicitly matched below
    "error",
    # matplotlib
    "ignore:Deprecated call to `pkg_resources\\.declare_namespace\\('.*'\\):DeprecationWarning",
    # ultralytics
    "ignore:pkg_resources is deprecated as an API",
    "ignore:Please use `gaussian_filter1d` from the `scipy.ndimage` namespace, the `scipy.ndimage.filters` namespace is deprecated.:DeprecationWarning",
]
